Split into separate files and fix BeginEdit() being called many times
