+++ src/AudioBand/AudioBand.csproj
@@ -101,6 +101,7 @@
     <Compile Include="Extensions\BindingExtensions.cs" />
     <Compile Include="Extensions\ColorExtensions.cs" />
     <Compile Include="Extensions\ImageExtensions.cs" />
+    <Compile Include="Extensions\ObjectExtensions.cs" />
     <Compile Include="Extensions\SvgExtensions.cs" />
     <Compile Include="MainControl.Bindings.cs">
       <SubType>UserControl</SubType>
